requirements.txt
python-telegram-bot==20.7
groq>=0.9.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.1
openai>=1.0.0
gTTS==2.5.1
pdfplumber==0.11.0
pydub==0.25.1
